Phase 5.4: RT-ASN1-TEMPLATE, and the two defects it found
The template interpreter landed in D81 with a stated hole: nothing drove CHOICE,
SEQUENCE, the SEQUENCE OF/SET OF content writer, the EMBED indirection or the
OPTIONAL absent answer through a caller-built descriptor, because no built-in item
has those shapes. The court now exists: a probe that declares its own structures
and its own ASN1_SEQUENCE and ASN1_CHOICE templates and drives them through the
public ASN1_item_* entry points, plus the twelve primitive-hook items and the two
ASN1_TYPE octet-string pairs. 100 observations.

It found two real defects:

  uint32_i2c negated at 64-bit width. The authority negates a uint32_t AS a
  uint32_t and only then widens; this crate widened first, so -5 encoded as nine
  octets (02 09 ff 00 00 00 00 ff ff ff fb) where the authority writes one
  (02 01 fb), and the candidate could not decode its own output.

  The SEQUENCE arm raised ASN1_R_FIELD_MISSING on a field error. The authority's
  'if (!ret) { errtt = seqtt; goto err; }' names the field and raises nothing
  further; treating it as 'the field was missing' put a fifth entry on the queue
  where the authority leaves four. The field_error flag is gone.

Three instrument and probe defects were fixed on the way in: the probe corrupted
the wrong byte in its wrong-tag case (index 2 held the value it wrote), drain
printed only the masked reason code, and run_courts.sh listed its runtime courts
by name so the new court produced a manifest and no receipt. The list is derived
from forensics/frf/courts/ now.

All gates green: 37 courts, 9537 observations; implemented[libcrypto] 865;
phase5 open 93. D83 records it.

+++ src/asn1/x_int64.rs
@@ -263,7 +263,7 @@ unsafe extern "C" fn uint32_i2c(
     it: *const Asn1Item,
 ) -> c_int {
     // SAFETY: the caller's contract.
-    let mut utmp = u64::from(unsafe { core::ptr::read_unaligned((*pval).cast::<u32>()) });
+    let mut utmp = unsafe { core::ptr::read_unaligned((*pval).cast::<u32>()) };
     let mut neg = 0;
     // SAFETY: `it` is a live item.
     let flags = unsafe { (*it).size };
@@ -271,13 +271,18 @@ unsafe extern "C" fn uint32_i2c(
     if flags & INTXX_FLAG_ZERO_DEFAULT == INTXX_FLAG_ZERO_DEFAULT && utmp == 0 {
         return -1;
     }
-    if flags & INTXX_FLAG_SIGNED == INTXX_FLAG_SIGNED && (utmp as u32 as i32) < 0 {
-        utmp = 0u64.wrapping_sub(utmp);
+    if flags & INTXX_FLAG_SIGNED == INTXX_FLAG_SIGNED && (utmp as i32) < 0 {
+        // The negation is at **32-bit** width, which is the whole point of this hook
+        // existing beside the 64-bit one: `0 - 0xfffffffb` is 5 in a `uint32_t` and
+        // 0xffffffff00000005 in a `uint64_t`. Widening first and negating second produced
+        // the second, which encoded `-5` as nine octets instead of one. `RT-ASN1-TEMPLATE`
+        // is what found it.
+        utmp = 0u32.wrapping_sub(utmp);
         neg = 1;
     }
 
     // SAFETY: `cont` is null or the caller's destination.
-    unsafe { ossl_i2c_uint64_int(cont, utmp, neg) }
+    unsafe { ossl_i2c_uint64_int(cont, u64::from(utmp), neg) }
 }
 
 /// `int uint32_c2i(ASN1_VALUE **pval, const unsigned char *cont, int len, int utype,
